+++ stat_linux.go
@@ -11,7 +11,7 @@ import (
 	"golang.org/x/sys/unix"
 )
 
-const supported SupportedType = Links | ATime | BTime | CTime | UID | GID | Syslinks
+const supports supportsType = supportsLinks | supportsBTime | supportsCTime | supportsSymlinks
 
 const userIDSource UserIDSourceType = UserIDSourceIsNumeric